Do not wait if $NVR_CMD is GUI application
On Windows, it's more natural that Command Prompt does not wait for GUI
application to finish. Launching applications including nvim-qt, gvim,
notepad, etc. from Command Prompt immediately returns the control to the
shell after the application starts. That behavior is very important
because Command Prompt does not provide handy way to execute the program
in background (like appending `&` or suspending and `bg` on UNIX).
Currently nvr waits for the child process to finish regardless of the
child process type, so make it not to wait for the child process if it's
GUI application on Windows.

To determine the process is a GUI application or not, you can look at
the PE header of the executable (`.exe` file). This implementation
depends on the library `pefile` to do so.